Renamed ProfileInputDialogFragment to more generic TextInputDialogFragment
